backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
supabase>=2.3.0
postgrest>=0.13.2
google-cloud-vision>=3.7.0
google-cloud-aiplatform>=1.38.1
google-cloud-speech>=2.24.0
google-auth>=2.27.0
pandas>=2.2.0
numpy>=1.26.0
Pillow>=10.4.0
PyPDF2>=3.0.1
pdf2image>=1.16.3
httpx>=0.26.0
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-api-python-client>=2.111.0
python-docx>=1.1.0
openpyxl>=3.1.0
markdown>=3.6
beautifulsoup4>=4.12.3
PymuPDF>=1.24.0
